Fourth Commit
Changes:
•Finished writing the 'GenerateLogString' method.
The method generates a random LogString following the defined format in
the Interview Assigment document.

•made progress with writing the 'RunLogger' method, setting up the timed
loop for the load test duration.

•Removed faulty call to method 'GetLoggingRateDuringPhase' using the
'loadTestLenInMin' variable.

•'RunLogger' method now receives the log file name instead of a List
<string> containing the file.

Newely Added:

•Added 'CheckIfToAddLine' method.
The method receives a value that represents the odd to add a line and
returns true if the random generated number is within the range of 1 and
the odd;