Fifth Commit
Changes:

•Merged the method that checks for valid numeric input and the method
that checks for valid string input into one method that handles both
validations.

•Changed the load generation implementation so it works more accurately
in terms of the precision of logs generated per second.

Newely Added:

•Added 'LiranHPAssignmentProject.rar'
A .rar file containing the complete Interview Assignment Project.

•Added 'Logger.cs' class file.
The following class contains the Logger object and its methods.

•Added 'ProgramStatusPrinter.cs' class file.
The class contains UI methods that display the program status to the
user.

•Added 'LiranHPAssignment.csproj' Visual Studion Project File.
The following project file manages and holds references to all the
objects who are part of the main Logger program.